#687: Better spacing in the UI, new prop for Dialog primary button that allows setting button variant.
