+++ new-client/src/components/Dialog/Dialog.js
@@ -30,6 +30,7 @@ export default function ResponsiveDialog(props) {
       allowDangerousHtml, // ReactMarkdown disables HTML by default but we let the Admin decide
       buttonText,
       headerText,
+      primaryButtonVariant,
       prompt,
       text,
       useLegacyNonMarkdownRenderer, // Admin can choose to pass-by the ReactMarkdown and just use dangerouslySetInnerHtml
@@ -97,7 +98,9 @@ export default function ResponsiveDialog(props) {
         )}
       </DialogContent>
       <DialogActions>
-        <Button onClick={handleClose}>{buttonText}</Button>
+        <Button onClick={handleClose} variant={primaryButtonVariant || "text"}>
+          {buttonText}
+        </Button>
         {abortText && <Button onClick={handleAbort}>{abortText}</Button>}
       </DialogActions>
     </Dialog>
